requirements.txt
# SquadSense Dependencies
# =======================

# --- Orchestration (Google ADK) ---
google-adk>=1.0.0
google-generativeai>=0.8.0

# --- LLM Providers ---
anthropic>=0.39.0

# --- Vector Database ---
chromadb>=0.5.0

# --- Code Parsing ---
tree-sitter>=0.23.0
tree-sitter-language-pack>=0.6.0

# --- Backend API ---
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# --- Frontend UI ---
streamlit>=1.40.0

# --- Data & Validation ---
pydantic>=2.0.0
python-dotenv>=1.0.0

# --- HTTP Client ---
httpx>=0.27.0
